+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fixed Ondo campaign portfolio swap account picker not recognising token balances when token addresses are stored in checksummed (EIP-55) form
+
 ## [7.72.1]
 
 ### Fixed